+++ src/features/auth/apis/index.ts
@@ -3,6 +3,13 @@ import {
   onErrorResponse,
   onResponse,
 } from '@/features/common/apis/interceptors';
+import {
+  AuthToken,
+  LogoutResponse,
+  UserInfo,
+  GoogleLoginUrl,
+} from '@/features/auth/apis/interfaces.ts';
+import { REQUEST_BODY_TYPE } from '@/constants/auth';
 
 export const oauthApi = axios.create({
   baseURL: 'https://www.googleapis.com',
@@ -19,3 +26,34 @@ setTimeout(() => {
 
 export const DOMAIN_URI = process.env.NEXT_PUBLIC_DOMAIN_URI;
 export const AUTH_KEY = 'AUTH' as const;
+
+export const getAuthCode = async (): Promise<GoogleLoginUrl> =>
+  await axios.get(`${DOMAIN_URI}/api/auth`).then(res => res.data);
+
+export const getAccessToken = async (code: string): Promise<AuthToken> =>
+  await axios
+    .post(`${DOMAIN_URI}/api/auth`, {
+      type: REQUEST_BODY_TYPE.GET_TOKEN,
+      code,
+    })
+    .then(res => res.data);
+
+export const getAccessTokenByRefreshToken = async (): Promise<AuthToken> =>
+  await axios
+    .post(`${DOMAIN_URI}/api/auth`, {
+      type: REQUEST_BODY_TYPE.GET_TOKEN_BY_REFRESH_TOKEN,
+      accessToken: localStorage.getItem('accessToken'),
+    })
+    .then(res => res.data);
+
+export const logout = async (
+  accessToken: string | null
+): Promise<LogoutResponse> =>
+  await axios
+    .delete(`${DOMAIN_URI}/api/auth?access_token=${accessToken}`)
+    .then(res => res.data);
+
+export const getUserInfo = async (accessToken: string): Promise<UserInfo> =>
+  await oauthApi
+    .get(`oauth2/v2/userinfo?access_token=${accessToken}`)
+    .then(res => res.data);
