+++ src/features/auth/apis/mutations/useGetAccessToken.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
+import { getAccessToken } from '@/features/auth/apis/index.ts';
+
+const useGetAccessToken = () =>
+  useMutation({
+    mutationFn: (code: string) => getAccessToken(code),
+  });
+
+export default useGetAccessToken;
